Reduce cross-version redundancy
