+++ src/spyglass/spikesorting/v0/spikesorting_burst.py
@@ -4,7 +4,6 @@ from typing import Dict, List, Tuple
 import datajoint as dj
 import matplotlib.pyplot as plt
 import numpy as np
-from datajoint.expression import QueryExpression
 from scipy import stats
 from spikeinterface.postprocessing.correlograms import (
     WaveformExtractor,
@@ -12,21 +11,28 @@ from spikeinterface.postprocessing.correlograms import (
 )
 
 from spyglass.decoding.utils import _get_peak_amplitude
+from spyglass.spikesorting.utils import (
+    calculate_ca,
+    calculate_isi_violation,
+    plot_burst_by_sort_group,
+    plot_burst_metrics,
+    plot_burst_pair_peaks,
+    plot_burst_peak_over_time,
+    plot_burst_xcorrel,
+    validate_pairs,
+)
 from spyglass.spikesorting.v0.spikesorting_curation import (
     CuratedSpikeSorting,
-    CuratedSpikeSortingSelection,
     Curation,
     Waveforms,
-    WaveformSelection,
 )
-from spyglass.spikesorting.v0.spikesorting_sorting import SpikeSorting
-from spyglass.utils import logger
+from spyglass.utils import SpyglassMixin, logger
 
 schema = dj.schema("burst_v0")  # TODO: rename to spikesorting_burst
 
 
 @schema
-class BurstPairParams(dj.Lookup):
+class BurstPairParams(SpyglassMixin, dj.Lookup):
     """Parameters for burst pair selection
 
     burst_params_name: name of the parameter set
@@ -56,19 +62,11 @@ class BurstPairParams(dj.Lookup):
 
     def get_params(self, key: dict) -> dict:
         """Given a key with burst_params_name, return the parameters"""
-        pk = self.primary_key[0]
-        if isinstance(key, str):
-            key = {pk: key}
-        if not isinstance(key, dict):
-            raise ValueError("key must be a dictionary")
-        passed_key = key.get(pk, None)
-        if not passed_key:
-            logger.warning("No key passed, using default")
-        return (self & {pk: passed_key or "default"}).fetch1("params")
+        return self.get_params_blob_from_key(key)
 
 
 @schema
-class BurstPairSelection(dj.Manual):
+class BurstPairSelection(SpyglassMixin, dj.Manual):
     definition = """
     -> CuratedSpikeSorting
     -> BurstPairParams
@@ -109,7 +107,9 @@ class BurstPairSelection(dj.Manual):
         }
 
         if sort_group_ids:  # restrict by passed sort_group_ids
-            query &= f'sort_group_id IN ({",".join(map(str, sort_group_ids))})'
+            query &= self.restrict_by_list(
+                "sort_group_id", sort_group_ids, return_restr=True
+            )
 
         # Skip duplicates unless specified otherwise
         kwargs["skip_duplicates"] = kwargs.get("skip_duplicates", True)
@@ -123,7 +123,7 @@ class BurstPairSelection(dj.Manual):
 
 
 @schema
-class BurstPair(dj.Computed):
+class BurstPair(SpyglassMixin, dj.Computed):
     definition = """
     -> BurstPairSelection
     """
@@ -224,38 +224,6 @@ class BurstPair(dj.Computed):
 
         return peak_amps, peak_timestamps
 
-    @staticmethod
-    def calculate_ca(bins: np.ndarray, correl: np.ndarray) -> float:
-        """Calculate Correlogram Asymmetry (CA)
-
-        defined as the contrast ratio of the area of the correlogram right and
-        left of coincident activity (zero).
-        http://www.psy.vanderbilt.edu/faculty/roeaw/edgeinduction/Fig-W6.htm
-
-        Parameters
-        ----------
-        bins : np.ndarray
-            array of bin edges
-        correl : np.ndarray
-            array of correlogram values
-        """
-        if not len(bins) == len(correl):
-            raise ValueError("Mismatch in lengths for correl asymmetry")
-        right = np.sum(correl[bins > 0])
-        left = np.sum(correl[bins < 0])
-        return 0 if (right + left) == 0 else (right - left) / (right + left)
-
-    @staticmethod
-    def calculate_isi_violation(
-        peak1: np.ndarray, peak2: np.ndarray, isi_threshold_s: float = 1.5
-    ) -> float:
-        """Calculate isi violation between two spike trains"""
-        spike_train = np.sort(np.concatenate((peak1, peak2)))
-        isis = np.diff(spike_train)
-        num_spikes = len(spike_train)
-        num_violations = np.sum(isis < (isi_threshold_s * 1e-3))
-        return num_violations / num_spikes
-
     def _compute_correlograms(
         self, key: dict, params: dict = None
     ) -> Tuple[np.ndarray, np.ndarray]:
@@ -304,9 +272,10 @@ class BurstPair(dj.Computed):
             return
 
         # mean waveforms in a dict: each one is of spike number x 4
+        waves = self._get_waves(key)
         waves_mean_1d = {
             u: np.reshape(
-                np.mean(self._get_waves(key).get_waveforms(u), axis=0).T,
+                np.mean(waves.get_waveforms(u), axis=0).T,
                 (1, -1),
             ).ravel()
             for u in units
@@ -325,10 +294,10 @@ class BurstPair(dj.Computed):
                     "wf_similarity": stats.pearsonr(
                         waves_mean_1d[u1], waves_mean_1d[u2]
                     ).statistic,
-                    "isi_violation": self.calculate_isi_violation(
+                    "isi_violation": calculate_isi_violation(
                         peak_timestamps[u1], peak_timestamps[u2]
                     ),
-                    "xcorrel_asymm": self.calculate_ca(
+                    "xcorrel_asymm": calculate_ca(
                         bins[1:], ccgs[u1 - 1, u2 - 1, :]
                     ),
                 }
@@ -337,40 +306,6 @@ class BurstPair(dj.Computed):
         self.insert1(key)
         self.BurstPairUnit.insert(unit_pairs)
 
-    @staticmethod
-    def _plot_metrics(sg_query):
-        """parameters are 4 metrics to be plotted against each other.
-
-        Parameters
-        ----------
-        wf_similarity : dict
-            waveform similarities
-        isi_violation : dict
-            isi violation
-        xcorrel_asymm : dict
-            spike cross correlogram asymmetry
-
-        Returns
-        -------
-        figure for plotting later
-        """
-
-        fig, ax = plt.subplots(1, 1, figsize=(12, 5))
-
-        for color_ind, row in enumerate(sg_query):
-            color = dict(color=f"C{color_ind}")
-            wf = row["wf_similarity"]
-            ca = row["xcorrel_asymm"]
-            ax.scatter(wf, ca, **color)
-            ax.text(wf, ca, f"({row['unit1']},{row['unit2']})", **color)
-
-        ax.set_xlabel("waveform similarity")
-        ax.set_ylabel("cross-correlogram asymmetry")
-
-        plt.close()
-
-        return fig
-
     def _get_fig_by_sg_id(self, key, sort_group_ids=None):
         query = self.BurstPairUnit & key
 
@@ -378,121 +313,40 @@ class BurstPair(dj.Computed):
             sort_group_ids = [sort_group_ids]
 
         if sort_group_ids:
-            query &= f'sort_group_id IN ({",".join(map(str, sort_group_ids))})'
+            query &= self.restrict_by_list(
+                "sort_group_id", sort_group_ids, return_restr=True
+            )
         else:
             sort_group_ids = np.unique(query.fetch("sort_group_id"))
 
         fig = {}
         for sort_group_id in sort_group_ids:
             sg_query = query & {"sort_group_id": sort_group_id}
-            fig[sort_group_id] = self._plot_metrics(sg_query)
+            fig[sort_group_id] = plot_burst_metrics(sg_query)
         return fig
 
-    def plot_by_sort_group_ids(self, key, sort_group_ids=None):
+    def plot_by_sort_group_ids(
+        self, key: dict, sort_group_ids: List[int] = None
+    ):
         fig = self._get_fig_by_sg_id(key, sort_group_ids)
-        for sg_id, f in fig.items():
-            title = f"sort group {sg_id}"
-            managed_fig, _ = plt.subplots(
-                1, 4, figsize=(24, 4), sharex=False, sharey=False
-            )
-            canvas_manager = managed_fig.canvas.manager
-            canvas_manager.canvas.figure = f
-            managed_fig.set_canvas(canvas_manager.canvas)
-            plt.suptitle(f"sort group {sg_id}", fontsize=20)
+        plot_burst_by_sort_group(fig)
 
-    def _validate_pair(
-        self, query: QueryExpression, unit1: int, unit2: int
-    ) -> Tuple[int, int]:
-        """Ensure that unit1, unit2 is a valid pair in the table.
-
-        Parameters
-        ----------
-        query : dj.QueryExpression
-            query to check for the pair. Subset of BurstPairUnit
-        unit1 : int
-            unit1 to check
-        unit2 : int
-            unit2 to check
-        """
-        if query & f"unit1={unit2} AND unit2={unit1}":
-            return unit1, unit2
-        elif query & f"unit1={unit1} AND unit2={unit2}":
-            logger.warning(f"Using reverse pair {unit1}, {unit2}")
-            return unit2, unit1
-        else:
-            logger.warning(f"No entry found for pair {unit1}, {unit2}")
-            return None
-
-    def _validate_pairs(self, key, pairs):
+    def investigate_pair_xcorrel(
+        self, key: dict, to_investigate_pairs: List[Tuple[int, int]]
+    ):
         query = self.BurstPairUnit & key
-        if isinstance(pairs, tuple) and len(pairs) == 2:
-            pairs = [pairs]
-        valid_pairs = []
-        for p in pairs:
-            if valid_pair := self._validate_pair(query, *p):
-                valid_pairs.append(valid_pair)
-        if not valid_pairs:
-            raise ValueError("No valid pairs found")
-        return valid_pairs
-
-    def investigate_pair_xcorrel(self, key, to_investigate_pairs):
-        used_pairs = self._validate_pairs(key, to_investigate_pairs)
-
-        col_num = int(np.ceil(len(used_pairs) / 2))
-
+        used_pairs = validate_pairs(query, to_investigate_pairs)
         fig = self._get_fig_by_sg_id(key)
-
-        fig, axes = plt.subplots(
-            2,
-            int(np.ceil(len(to_investigate_pairs) / 2)),
-            figsize=(col_num * 3, 4),
-            squeeze=True,
-        )
-
         ccgs_e, bins = self._compute_correlograms(key)
+        plot_burst_xcorrel(fig, ccgs_e, bins, used_pairs)
 
-        for ind, p in enumerate(used_pairs):
-            (u1, u2) = p
-            axes[np.unravel_index(ind, axes.shape)].bar(
-                bins[1:], ccgs_e[u1 - 1, u2 - 1, :]
-            )
-            axes[np.unravel_index(ind, axes.shape)].set_xlabel("ms")
-
-        if len(used_pairs) < col_num * 2:  # remove the last unused axis
-            axes[np.unravel_index(ind, axes.shape)].axis("off")
-
-        plt.tight_layout()
-
-    def investigate_pair_peaks(self, key, to_investigate_pairs):
-        used_pairs = self._validate_pairs(key, to_investigate_pairs)
+    def investigate_pair_peaks(
+        self, key: dict, to_investigate_pairs: List[Tuple[int, int]]
+    ):
+        query = self.BurstPairUnit & key
+        used_pairs = validate_pairs(query, to_investigate_pairs)
         peak_amps, peak_timestamps = self.get_peak_amps(key)
-
-        fig, axes = plt.subplots(
-            len(used_pairs), 4, figsize=(12, 2 * len(used_pairs))
-        )
-
-        def get_kwargs(unit, data):
-            return dict(
-                alpha=0.5,
-                weights=np.ones(len(data)) / len(data),
-                label=str(unit),
-            )
-
-        for ind, (u1, u2) in enumerate(used_pairs):
-
-            peak1 = peak_amps[u1]
-            peak2 = peak_amps[u2]
-
-            axes[ind, 0].set_ylabel("percent")
-            for i in range(4):
-                data1, data2 = peak1[:, i], peak2[:, i]
-                axes[ind, i].hist(data1, **get_kwargs(u1, data1))
-                axes[ind, i].hist(data2, **get_kwargs(u2, data2))
-                axes[ind, i].set_title("channel " + str(i + 1))
-                axes[ind, i].set_xlabel("uV")
-                axes[ind, i].legend()
-
-        plt.tight_layout()
+        plot_burst_pair_peaks(used_pairs, peak_amps, peak_timestamps)
 
     def plot_peak_over_time(
         self,
@@ -511,62 +365,7 @@ class BurstPair(dj.Computed):
         overlap : bool, optional
             if True, plot units in pair on the same plot
         """
-
         peak_v, peak_t = self.get_peak_amps(key)
-
-        for pair in self._validate_pairs(key, to_investigate_pairs):
-            kwargs = (
-                dict(fig=None, axes=None, row_duration=100)
-                if overlap
-                else dict()
-            )
-
-            for u in pair:
-                ret1, ret2 = self.plot_1peak_over_time(
-                    peak_v[u], peak_t[u], show_plot=overlap, **kwargs
-                )
-                if overlap:
-                    fig, axes = ret1, ret2
-                    kwargs = dict(fig=fig, axes=axes)
-                else:
-                    if fig is None:
-                        fig, axes = dict(), dict()
-                    fig[u], axes[u] = ret1, ret2
-
-            axes[0, 0].set_title(f"pair:{pair}")
-
-    def plot_1peak_over_time(
-        self,
-        voltages,
-        timestamps,
-        fig: plt.Figure = None,
-        axes: plt.Axes = None,
-        row_duration: int = 600,
-        show_plot: bool = False,
-    ):
-        max_channel = np.argmax(-np.mean(voltages, 0))
-        time_since = timestamps - timestamps[0]
-        row_num = int(np.ceil(time_since[-1] / row_duration))
-
-        if axes is None:
-            fig, axes = plt.subplots(
-                row_num,
-                1,
-                figsize=(20, 2 * row_num),
-                sharex=True,
-                sharey=True,
-                squeeze=False,
-            )
-
-        for ind in range(row_num):
-            t0 = ind * row_duration
-            t1 = t0 + row_duration
-            sub_ind = np.logical_and(time_since >= t0, time_since <= t1)
-            axes[ind, 0].scatter(
-                time_since[sub_ind] - t0, voltages[sub_ind, max_channel]
-            )
-
-        if not show_plot:
-            plt.close()
-
-        return fig, axes
+        query = self.BurstPairUnit & key
+        used_pairs = validate_pairs(query, to_investigate_pairs)
+        plot_burst_peak_over_time(peak_v, peak_t, used_pairs, overlap=overlap)
